Writing and regeneration of the report and some figures.
I include in the regression part finding the lowest value of the RMS
error, because I had forgotten to do this. This value is included in the
report. I expand the discussion of the k-NN results, I think we should
do a plot of the knollC data when applying M to all the data points.