Finalise report, README and package script.
We finalise the report, make comparisons between LDA and k-NN and include a visualisation of
knollC test set before and after transforming it using the d metric (in useknn.m)

We introduce a package script, to pack everything for the _actual_ handin.

In lda.m we print the mean values and the common sigma.