Added report to repo, modified SVM model selection.
The modelselect method now returns the configuration for C and gamma
that give the best results (using 5-fold cross validation). These are
then passed to the train method, and the resultant graph seems lighter
in terms of support vectors (there are _way_ fewer).

runsvm for some reason only passed the entire data set (as one) to the
train function, whereas it seems to expect the data in two vectors - one
for data and one for labels. I modified the calls in runsvm to do this.